dfgr

+++ excelBiulderTest.groovy
@@ -73,7 +73,7 @@ headerRow.setHeight((short) (23*20));  // Setzen Sie die Zeilenhöhe auf 19
         XSSFCell cell = headerRow.createCell(i);
         cell.setCellValue(headersHeadline[i]);
         cell.setCellStyle(headerstyle);
-        cell.setEncoding(HSSFCell.ENCODING_UTF_16);
+        cell.setEncoding((short) HSSFCell.ENCODING_UTF_16);
             
     }
 
@@ -149,7 +149,7 @@ for (int rowNum = 0; rowNum < data.length; rowNum++) {
             XSSFCell cell = dataRow.createCell(colNum);
             cell.setCellValue(data[rowNum][colNum]);
             cell.setCellStyle(style);
-            cell.setEncoding(HSSFCell.ENCODING_UTF_16);
+            cell.setEncoding((short) HSSFCell.ENCODING_UTF_16);
         }
 
     }else{
@@ -158,7 +158,7 @@ for (int rowNum = 0; rowNum < data.length; rowNum++) {
             XSSFCell cell = dataRow.createCell(colNum);
             cell.setCellValue(data[rowNum][colNum]);
             cell.setCellStyle(style1);
-            cell.setEncoding(HSSFCell.ENCODING_UTF_16);
+            cell.setEncoding((short) HSSFCell.ENCODING_UTF_16);
         }
 
     }
